refactor(#184): 채팅방 관련 기능 ChatRoomServiceImpl로 옮김
