+++ src/main/java/com/cotato/kampus/domain/chat/api/port/ChatRoomService.java
@@ -1,13 +1,13 @@
 package com.cotato.kampus.domain.chat.api.port;
 
 import com.cotato.kampus.domain.chat.domain.ChatRoomDetailDto;
-import com.cotato.kampus.domain.chat.domain.ChatRoomPreview;
+import com.cotato.kampus.domain.chat.domain.ChatRoomPreviewList;
 
 public interface ChatRoomService {
 
-	public Long createChatRoom(Long id);
+	public Long createChatRoom(Long postId);
 
-	public ChatRoomPreview findChatRooms(int page);
+	public ChatRoomPreviewList findChatRooms(int page);
 
 	public ChatRoomDetailDto getChatRoomDetail(Long chatroomId);
