+++ src/main/java/com/cotato/kampus/domain/chat/application/ChatRoomServiceImpl.java
@@ -1,0 +1,142 @@
+package com.cotato.kampus.domain.chat.application;
+
+import java.util.List;
+
+import org.springframework.data.domain.Slice;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.cotato.kampus.domain.board.application.BoardFinder;
+import com.cotato.kampus.domain.board.dto.BoardDto;
+import com.cotato.kampus.domain.chat.api.port.ChatRoomService;
+import com.cotato.kampus.domain.chat.dao.entity.ChatRoomEntity;
+import com.cotato.kampus.domain.chat.dao.entity.ChatroomMetadataEntity;
+import com.cotato.kampus.domain.chat.domain.ChatRoomDetailDto;
+import com.cotato.kampus.domain.chat.domain.ChatRoomPreview;
+import com.cotato.kampus.domain.chat.domain.ChatRoomPreviewList;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomAppender;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomDeleter;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomFinder;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomValidator;
+import com.cotato.kampus.domain.chat.implement.message.ChatMessageDeleter;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataAppender;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataDeleter;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataFinder;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataMapper;
+import com.cotato.kampus.domain.chat.implement.read.MessageReadStatusDeleter;
+import com.cotato.kampus.domain.common.application.ApiUserResolver;
+import com.cotato.kampus.domain.post.application.PostFinder;
+import com.cotato.kampus.domain.post.dto.PostDto;
+import com.cotato.kampus.domain.post.dto.PostReferenceDto;
+
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
+@Slf4j
+public class ChatRoomServiceImpl implements ChatRoomService {
+
+	private final ChatRoomValidator chatRoomValidator;
+	private final ChatRoomAppender chatRoomAppender;
+	private final ChatRoomFinder chatRoomFinder;
+	private final ChatRoomDeleter chatroomDeleter;
+
+	private final ApiUserResolver apiUserResolver;
+
+	private final PostFinder postFinder;
+	private final BoardFinder boardFinder;
+
+	private final ChatMessageDeleter chatMessageDeleter;
+	private final MessageReadStatusDeleter messageReadStatusDeleter;
+
+	private final ChatroomMetadataAppender chatroomMetadataAppender;
+	private final ChatroomMetadataFinder chatroomMetadataFinder;
+	private final ChatroomMetadataMapper chatroomMetadataMapper;
+	private final ChatroomMetadataDeleter chatroomMetadataDeleter;
+
+	@Override
+	public Long createChatRoom(Long postId) {
+		// 1. 게시글 정보 조회
+		PostDto post = postFinder.findPost(postId);
+		Long receiverId = post.userId();
+		String postTitle = post.title();
+
+		// 2. 채팅을 건 유저를 조회
+		Long senderId = apiUserResolver.getCurrentUserId();
+
+		// 3. 채팅방 생성 검증(보내는 사람 == 받는 사람 or 이미 있는 채팅방)
+		chatRoomValidator.validateNewChatRoom(postId, senderId, receiverId);
+
+		// 4. 채팅방 생성
+		Long chatroomId = chatRoomAppender.appendChatRoom(postId, senderId, receiverId);
+
+		// 5. 채팅방 리스트 조회시 사용되는 뷰 생성
+		chatroomMetadataAppender.appendChatroomMetadatas(
+			chatroomId,
+			postId,
+			postTitle,
+			senderId,
+			receiverId
+		);
+
+		return chatroomId;
+	}
+
+	@Override
+	public ChatRoomPreviewList findChatRooms(int page) {
+		// 1. 유저 정보를 조회
+		Long userId = apiUserResolver.getCurrentUserId();
+
+		// 2. 해당 유저의 채팅방 메타데이터를 lastChatTime 내림차순으로 조회
+		Slice<ChatroomMetadataEntity> chatRoomMetadatas = chatroomMetadataFinder.findChatRoomMetadatas(userId, page);
+
+		// 3. ChatRoomPreview로 변환
+		List<ChatRoomPreview> previewList = chatRoomMetadatas.getContent()
+			.stream()
+			.map(chatroomMetadataMapper::toChatRoomPreview)
+			.toList();
+
+		return ChatRoomPreviewList.from(previewList, chatRoomMetadatas.hasNext());
+	}
+
+	@Override
+	public ChatRoomDetailDto getChatRoomDetail(Long chatroomId) {
+		// 1. Find chatroom
+		ChatRoomEntity chatRoomEntity = chatRoomFinder.findChatroom(chatroomId);
+		// 2. 게시글 정보 가져옴
+		PostReferenceDto postReference = postFinder.findPostReference(chatRoomEntity.getPostId());
+
+		// 3. 게시글이 삭제된 경우
+		if (postReference.isDeleted()) {
+			return ChatRoomDetailDto.ofDeleted(chatRoomEntity, postReference);
+		}
+		// 4. 게시글이 존재하는 경우
+		BoardDto board = boardFinder.findBoardDto(postReference.boardId());
+		return ChatRoomDetailDto.of(chatRoomEntity, postReference, board);
+	}
+
+	@Override
+	@Transactional
+	public void deleteChatroom(Long chatroomId) {
+		// 1. 현재 사용자 ID 조회
+		Long userId = apiUserResolver.getCurrentUserId();
+
+		// 2. 채팅방 멤버 검증
+		chatRoomValidator.validateUser(userId, chatroomId);
+
+		// 3. 채팅 메시지 삭제(sender, receiver)
+		chatMessageDeleter.deleteByChatroomId(chatroomId);
+
+		// 4. 읽음 상태 삭제(sender, receiver)
+		messageReadStatusDeleter.deleteByChatroomId(chatroomId);
+
+		// 5. 채팅방 메타데이터 삭제(sender, receiver)
+		chatroomMetadataDeleter.deleteByChatroomId(chatroomId);
+
+		// 6. 채팅방 삭제
+		chatroomDeleter.deleteById(chatroomId);
+	}
+}
